feat(dispatch): refactor handler return type interface
Previously, we had two distinct traits for message handlers:

* `Handler<T>` for handlers that only handled a `Message` and did not
  return anything back to the originator peer
* `ReplyHandler<T>` for handlers that wanted to return a response or
  error back to the originator peer

This interface was not very flexible because the `ReplyHandler<T>` was
enforcing a return of a Rust `Result<T, E>`, which meant that users
could not easily choose to only return a response or an error back to
the originator.

Also, the user had to use two distinct functions to register handlers to
the registry, e.g `handles::<T>` for unit handlers or `replies::<T>`
for reply handlers.

This commit introduces a new `Response` abstraction that is used to
represent a response returned by a handler, along with a `IntoResponse`
trait that can be implemented for types that can be converted to a
`Response`.

This allows us to completly remove `ReplyHandler` and keep only a single
`Handler<T>` trait which now exposes a new `Response` associated type
that the user can specify to return any type for which `IntoResponse` is
implemented