feat(core): embed prost as `zebus-proto-*` crates to make `prost::Message::encode` object-safe
prost exposes a `Message` trait to encode or decode a protobuf message.
The `Message` trait is object-safe but does not allow one to either encode or decode
a message through a `dyn Message` reference.

On a road to make our `Bus` trait object-safe and allow the user to call
`send_` methods through a `dyn Bus` reference as well, we need to remove
the generics from the `send*`() methods.

First step of removing generics consists of having a protobuf trait
implementation that allows to encoding a message through a `dyn Message` reference.

Since the original prost implementation does not allow encoding a
`Message` through a trait-object (see https://github.com/tokio-rs/prost/issues/742),
prost has been embedded as external crates
(`prost` as `zebus-proto` and `prost-derive` as `zebus-proto-derive`)
and `encode()` can now be invoked through `dyn Message` since it is not
generic anymore.

`prost::*` types and derive macros are re-exported in the `zebus::proto`
module to make it easy to migrate to `prost` again in the future if the
constraint on `Message::encode()` is relaxed