feat(core): make `Bus` trait object-safe
Prior to this commit, the `Bus` was not object-safe because `send_*`
functions were generic over the type of command to send.

This means that the `Bus` could not be used through a `&dyn Bus` or
`Arc<dyn Bus>`.

However, it is expected that users might want to inject a `Arc<dyn Bus>`
into their own components or even handlers to be able to send commands
or publish events through it.

This limitation has now been lifted by making `send_*` non-generic and
take `&dyn Command` as a parameter instead of being generic over a `C:
Command`.

Note that work is still needed to make `Bus` `Send` and `Sync` so that
users can wrap it inside an `Arc` and use it across threads