chore(persistence): properly await for the inner transport start completion future when starting the transport
When starting the PersistenceTransport, the inner transport was first started. The start() operation
is supposed to return a future that completes once the start sequence completes

However, when starting the persistence transport, the underlying transport was started but the future
it returned was not being awaited on. This was not causing any issue as start() is supposed to have
a side-effect even though the future it returns is not polled since it is a *COMPLETION* future and
not a future for the start operation itself

Even though in this particular case, it did not represent any issue because the underlying transport
(either zmq or memory) returned a future that was always ready, the compiler was still complaining
with a valid reason

We now return a future that properly awaits the inner transport future before making any progress

This commit also allows us to get rid of the special StartFuture and we now directly return
a BoxFuture of an async block from which we can directly call await on instead of maintaining
our own state machine and polling our own underlying futures